Add windows specific configurations
I've installed neovide on windows and might use for coding outside *nix
systems. The git bash terminal that I use is usually enough, but it
doesn't hurt to have more options.

This allows for using the git bash shell instead of powershell when
invoking ':terminal', a better font (as well as some alternatives), and
a setting to make the window fullscreen on windows.